+++ api/V0GetKubeconfig.go
@@ -1,8 +1,6 @@
 package api
 
 import (
-	"crypto/tls"
-	"io"
 	"net/http"
 	"strconv"
 	"strings"
@@ -21,43 +19,15 @@ func (e *API) V0GetKubeconfig(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	client := &http.Client{}
-	// #nosec G402
-	client.Transport = &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: e.Config.SkipSSL},
+	content := e.Redis.GetRedisKey(e.Framework.FrameworkName + ":kubernetes_config")
+
+	if content != "" {
+		// replace the localhost server string with the mesos agent hostname and dynamic port
+		destURL := e.Config.K3SServerHostname + ":" + strconv.Itoa(e.Config.K3SServerPort)
+		kubconf := strings.Replace(string(content), "127.0.0.1:6443", destURL, -1)
+
+		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+		w.Header().Set("Api-Service", "v0")
+		w.Write([]byte(kubconf))
 	}
-	req, _ := http.NewRequest("GET", e.BootstrapProtocol+"://"+e.Config.K3SServerHostname+":"+strconv.Itoa(e.Config.K3SServerContainerPort)+"/api/m3s/bootstrap/v0/config", nil)
-	req.SetBasicAuth(e.Config.BootstrapCredentials.Username, e.Config.BootstrapCredentials.Password)
-	req.Close = true
-	res, err := client.Do(req)
-
-	if err != nil {
-		logrus.WithField("func", "V0GetKubeconfig").Error(err.Error())
-		w.WriteHeader(http.StatusNotAcceptable)
-		return
-	}
-
-	defer res.Body.Close()
-
-	if res.StatusCode != 200 {
-		logrus.WithField("func", "V0GetKubeconfig").Error("Response Code: ", res.StatusCode)
-		w.WriteHeader(http.StatusNotAcceptable)
-		return
-	}
-
-	content, err := io.ReadAll(res.Body)
-
-	if err != nil {
-		logrus.WithField("func", "V0GetKubeconfig").Error(err.Error())
-		w.WriteHeader(http.StatusNotAcceptable)
-		return
-	}
-
-	// replace the localhost server string with the mesos agent hostname and dynamic port
-	destURL := e.Config.K3SServerHostname + ":" + strconv.Itoa(e.Config.K3SServerPort)
-	kubconf := strings.Replace(string(content), "127.0.0.1:6443", destURL, -1)
-
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-	w.Header().Set("Api-Service", "v0")
-	w.Write([]byte(kubconf))
 }